Update README.md
We came to the conclusion in discord that this step of installing ffmpeg using winget should not be missing since adding ffmpeg.exe to the dependencies folder doesnt work that easily. (or doesn't work at all anymore since there's no "SET "FFMPEG_PATH=%DEPENDENCIES%" " in the starting batch files from the original repository.